todos os endpoints implementados

+++ apps/igti-node-desafiofinal/src/autor/autor.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Prisma } from '@prisma/client';
 import { PrismaService } from '../app/prisma.service';
 
@@ -23,6 +27,20 @@ export class AutorService {
   }
 
   async remove(autor_id: number) {
-    return this.prisma.autores.delete({ where: { autor_id } });
+    const linkedBook = await this.prisma.livros.findFirst({
+      where: { autor_id },
+    });
+    if (linkedBook !== null) {
+      // TODO: create a bussiness exception
+      throw new BadRequestException(
+        'Não é permitida a deleção de autor com livro vinculado.'
+      );
+    } else {
+      try {
+        await this.prisma.autores.delete({ where: { autor_id } });
+      } catch (error) {
+        throw new NotFoundException();
+      }
+    }
   }
 }
